Reorganized some code (testing functions).
PriceBarChart unit conversions (x to/from timestamp and y to/from price) are now done from class PriceBarChartGraphicsScene.  

The datetimeToStr() function that was needed for a wider range of dated has now moved to the Ephemeris class as a static function.  

Added edit widget and dialog classes for editing a PriceBarChartBarCountArtifact in the context of  PriceBarChart.

All code changes not fully tested yet.